pyproject.toml
[tool.poetry]
name = "auth_sprint_1"
version = "0.1.0"
description = ""
authors = ["kenki <ilsafganiev2222@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask-JWT-Extended = "^4.4.4"
flask-marshmallow = "^0.14.0"
Flask-Security = "^3.0.0"
Flask-SQLAlchemy = "^2.5.1"
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
psycopg2-binary = "^2.9.3"
email-validator = "^1.2.1"
Flask = "2.1.2"
Werkzeug = "2.1.2"
bcrypt = "^4.0.0"
apispec-webframeworks = "^0.5.2"
marshmallow = "^3.17.1"
flask-swagger-ui = "^4.11.1"
gevent = "^21.12.0"
redis = "^4.3.4"
click = "^8.1.3"
Flask-Migrate = "^3.1.0"
rauth = "^0.7.3"

[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.6.22"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-variables-names = "^0.0.5"
pre-commit = "^2.19.0"
dlint = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
yesqa = "^1.4.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
aiohttp = "^3.8.1"

[tool.black]
line-length = 120
target-version = ["py38"]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.venv
    | \.git
    | \.env
  )/
)
'''

[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
